# docs/08-openapi.yaml
...
openapi: 3.1.0
info:
  title: TokenTrail API
  version: 1.0.0
  description: >
    Control-plane REST API for TokenTrail, the open-source AI cost governance and
    usage analytics platform. Gateway (data-plane) routes under /gw are documented
    separately per provider surface.
  license: { name: Apache-2.0, identifier: Apache-2.0 }
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Access token issued; refresh token set as httpOnly cookie.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  user: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Problem" }
...
  /workspaces/{ws}/invitations:
    parameters: [{ $ref: "#/components/parameters/ws" }]
    post:
      tags: [Workspaces]
      summary: Invite a user (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { $ref: "#/components/schemas/WorkspaceRole" }
                teamId: { type: string, format: uuid }
      responses:
        "201": { description: Invitation created and emailed }
...
  /workspaces/{ws}/credentials:
    parameters: [{ $ref: "#/components/parameters/ws" }]
    post:
      tags: [Credentials]
      summary: Add a provider credential (ADMIN). Secret is write-only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, name]
              properties:
                provider: { $ref: "#/components/schemas/Provider" }
                name: { type: string }
                secret: { type: string, writeOnly: true }
                baseUrl: { type: string, format: uri, description: "Required for OLLAMA; optional elsewhere" }
                modelAllowlist: { type: array, items: { type: string } }
                isDefault: { type: boolean }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Credential" } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: "JWT or ttp_ personal access token" }
  parameters:
    ws: { name: ws, in: path, required: true, schema: { type: string }, description: Workspace id or slug }
...
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        requestId: { type: string }
        errors:
          type: array
          items: { type: object, properties: { path: { type: string }, message: { type: string } } }
...
    WorkspaceRole: { type: string, enum: [OWNER, ADMIN, MEMBER, VIEWER] }
    Provider: { type: string, enum: [ANTHROPIC, OPENAI, GEMINI, MINIMAX, OPENROUTER, DEEPSEEK, OLLAMA] }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        status: { type: string, enum: [ACTIVE, DEACTIVATED] }
...
    Credential:
      type: object
      properties:
        id: { type: string, format: uuid }
        provider: { $ref: "#/components/schemas/Provider" }
        name: { type: string }
        secretLast4: { type: [string, "null"] }
        baseUrl: { type: [string, "null"] }
        isDefault: { type: boolean }
        status: { type: string, enum: [ACTIVE, DISABLED] }
...